+++ tests/integration/test_get_backend.py
@@ -50,7 +50,7 @@ def test_get_malformed_backend(api):
     """Raises TypeError if a malformed backend is returned"""
     provider = _get_test_provider(url=API_URL)
 
-    err_msg = "(TergiteBackendConfig.)?__init__\\(\\) missing 1 required positional argument: 'version'"
+    err_msg = "1 validation error for TergiteBackendConfig\nversion\n"
     with pytest.raises(
         TypeError, match=f"malformed backend '{MALFORMED_BACKEND}', {err_msg}"
     ):